Adjust namespaces for auth related packets and rename Constants to AuthConsts
